[router] Refactor router; Fix recaptcha not loaded in the homepage; Create separate file for routes.
