+++ src/router.js
@@ -1,16 +1,10 @@
-import setHomePage from "./pages/home";
-import setWorkPage from "./pages/work";
+import { routes } from './js/routes.js';
 
 export const route = (event) => {
    event = event || window.event;
    event.preventDefault();
    window.history.pushState({}, "", event.target.href);
    handleLocation();
-}
-
-const routes = {
-   '/': () => setHomePage(document.getElementById('main-content')),
-   '/work': () => setWorkPage(document.getElementById('main-content')),
 }
 
 const handleLocation = async () => {
